% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_floralShoppe.R, R/theme_jwz.R,
%   R/theme_newRetro.R
\name{check_font}
\alias{check_font}
\title{Check for fonts}
\usage{
check_font(font_name)

check_font(font_name)

check_font(font_name)
}
\arguments{
\item{font_name}{}

\item{font_name}{}

\item{font_name}{}
}
\description{
Check for fonts

Check for font

Check for font
}
