% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_crystalPepsi.R
\name{crystalPepsi_pal}
\alias{crystalPepsi_pal}
\title{Crystal Pepsi Colour Palette}
\usage{
crystalPepsi_pal()
}
\value{
A character vector of color hex codes.
}
\description{
Crystal Pepsi Colour Palette
}
