% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_floralShoppe.R
\name{floral_shoppe}
\alias{floral_shoppe}
\title{Floral Shoppe Theme}
\usage{
floral_shoppe(
  font = "VCR OSD Mono",
  main.text.color = "#6CFFB9",
  sub.text.color = "#6CFFB9",
  base.size = 15,
  plot.background.color = "#FF819C",
  legend.position = "right",
  panel.border.color = "black",
  panel.background = "black",
  panel.grid.color = "#6CFFB9",
  axis.text.color = "#6CFFB9",
  axis.text.size = base.size * 0.6666667,
  subtitle.size = base.size * 0.8,
  legend.text = base.size * 0.8,
  legend.title = base.size * 0.93,
  axis.title.size = base.size * 0.8,
  title.size = 15
)
}
\arguments{
\item{font}{Base text family}

\item{main.text.color}{Color of main text}

\item{sub.text.color}{Color of smaller text}

\item{base.size}{Base text size}

\item{plot.background.color}{Color of plot background, passed to `plot.background`}

\item{legend.position}{Position of legend}

\item{panel.border.color}{Color of plot panel border}

\item{panel.background}{Background panel color}

\item{panel.grid.color}{Color of the panel grid}

\item{axis.text.color}{Color of axis text}

\item{axis.text.size}{Size of axis text}

\item{subtitle.size}{Size of subtitle}

\item{legend.text}{Size of legend}

\item{legend.title}{Size of legend title}

\item{axis.title.size}{Size of axis title}

\item{title.size}{Size of title}
}
\description{
ggplot2 plot theme based on the Floral Shoppe album cover
}
\references{
https://beerontherug.bandcamp.com/album/floral-shoppe
}
\seealso{
[ggplot2::theme]
}
