% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_floralShoppe.R
\name{floralShoppe_pal}
\alias{floralShoppe_pal}
\title{Floral Shoppe Colour Palette}
\usage{
floralShoppe_pal()
}
\value{
A character vector of color hex codes.
}
\description{
Floral Shoppe Colour Palette
}
