% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hotlineBling.R
\name{hotlineBling_pal}
\alias{hotlineBling_pal}
\title{Hotline Bling Colour Palette}
\usage{
hotlineBling_pal()
}
\value{
A character vector of color hex codes.
}
\description{
Hotline Bling Colour Palette
}
