% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hyperBubble.R
\name{hyperBubble_pal}
\alias{hyperBubble_pal}
\title{Hyper Bubble Colour Palette}
\usage{
hyperBubble_pal()
}
\value{
A character vector of color hex codes.
}
\description{
Hyper Bubble Colour Palette
}
