% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_jazzCup.R
\name{jazzCup_pal}
\alias{jazzCup_pal}
\title{Jazz Cup Colour Palette}
\usage{
jazzCup_pal()
}
\value{
A character vector of color hex codes representing the Jazz Cup palette.
}
\description{
Colour scales inspired by the iconic Solo Jazz pattern found on waxed paper cups in the '90s.
}
