% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jwz.R
\name{jwz}
\alias{jwz}
\title{jwz Theme}
\usage{
jwz(font = "Windows Command Prompt", text.color = "#02BAFF",
  base.size = 15, plot.background.color = "black",
  legend.position = "right", panel.border.color = "black",
  panel.background = "#014400", panel.grid.color = "#00FF01",
  axis.text.color = "#02BAFF", axis.text.size = base.size * 0.6666667,
  subtitle.size = base.size * 0.8, legend.text = base.size * 0.8,
  legend.title = base.size * 0.93, axis.title.size = base.size * 0.8,
  title.size = 15)
}
\arguments{
\item{font}{Base text family}

\item{text.color}{Color of text}

\item{base.size}{Base text size}

\item{plot.background.color}{Color of plot background, passed to `plot.background`}

\item{legend.position}{Position of legend}

\item{panel.border.color}{Color of plot panel border}

\item{panel.background}{Background panel color}

\item{panel.grid.color}{Color of the panel grid}

\item{axis.text.color}{Color of axis text}

\item{axis.text.size}{Size of axis text}

\item{subtitle.size}{Size of subtitle}

\item{legend.text}{Size of legend}

\item{legend.title}{Size of legend title}

\item{axis.title.size}{Size of axis title}

\item{title.size}{Size of title}
}
\description{
[ggplot2] plot theme based on the personal blog of Jamie Zawinski
}
\references{
https://en.wikipedia.org/wiki/Jamie_Zawinski
}
\seealso{
[ggplot2::theme]
}
