% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_jwz.R
\name{jwz_pal}
\alias{jwz_pal}
\title{jwz Colour Palette}
\usage{
jwz_pal()
}
\value{
A character vector of color hex codes representing the jwz palette.
}
\description{
Colour scales inspired by the personal blog of computer programmer Jamie Zawinski.
}
