% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_macPlus.R
\name{macPlus_pal}
\alias{macPlus_pal}
\title{Macintosh Plus Colour Palette}
\usage{
macPlus_pal()
}
\value{
A character vector of color hex codes representing the Macintosh Plus palette.
}
\description{
Colour scales inspired by the Macintosh Plus computer released in 1986.
}
