% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_mallSoft.R
\name{mallSoft_pal}
\alias{mallSoft_pal}
\title{Mallsoft Colour Palette}
\usage{
mallSoft_pal()
}
\value{
A character vector of color hex codes representing the Mallsoft palette.
}
\description{
Colour scales inspired by the Mallsoft subgenre of vaporwave music.
Mallsoft emphasizes the sense of consumeristic detachment in the ambience of a suburban shopping mall.
}
