% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_newRetro.R
\name{newRetro_pal}
\alias{newRetro_pal}
\title{New Retro Colour Palette}
\usage{
newRetro_pal()
}
\value{
A character vector of color hex codes representing the New Retro palette.
}
\description{
Colour scales inspired by the New Retro vaporwave record label "Live the 80's Dream!"
}
