% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_avanti.R
\name{scale_colour_avanti}
\alias{scale_colour_avanti}
\alias{scale_color_avanti}
\title{Avanti Colour and Fill Scales for ggplot2}
\usage{
scale_colour_avanti(...)

scale_color_avanti(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Avanti Colour and Fill Scales for ggplot2
}
\seealso{
\code{\link[ggplot2:scale_colour_discrete]{scale_colour_discrete()}} and 
  \code{\link[ggplot2:scale_fill_discrete]{scale_fill_discrete()}}

\code{\link{scale_colour_avanti}}
}
