% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_crystalPepsi.R
\name{scale_colour_crystalPepsi}
\alias{scale_colour_crystalPepsi}
\title{crystalPepsi colour}
\usage{
scale_colour_crystalPepsi(...)
}
\arguments{
\item{...}{pass the palette}
}
\value{
crystalPepsi colour
}
\description{
crystalPepsi colour
}
