% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_jazzCup.R
\name{scale_colour_jazzCup}
\alias{scale_colour_jazzCup}
\alias{scale_color_jazzCup}
\title{Jazz Cup Colour and Fill Scales for ggplot2}
\usage{
scale_colour_jazzCup(...)

scale_color_jazzCup(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Jazz Cup Colour and Fill Scales for ggplot2

Scale Color Jazz Cup - Alias for Colour
}
\seealso{
\code{\link[ggplot2:scale_colour_discrete]{scale_colour_discrete()}} and 
  \code{\link[ggplot2:scale_fill_discrete]{scale_fill_discrete()}}

\code{\link{scale_colour_jazzCup}}
}
