% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_macPlus.R
\name{scale_colour_macPlus}
\alias{scale_colour_macPlus}
\alias{scale_color_macPlus}
\title{Macintosh Plus Colour and Fill Scales for ggplot2}
\usage{
scale_colour_macPlus(...)

scale_color_macPlus(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Macintosh Plus Colour and Fill Scales for ggplot2

Scale Color Macintosh Plus - Alias for Colour
}
\seealso{
\code{\link[ggplot2:scale_colour_discrete]{scale_colour_discrete()}} and 
  \code{\link[ggplot2:scale_fill_discrete]{scale_fill_discrete()}}

\code{\link{scale_colour_macPlus}}
}
