% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_newRetro.R
\name{scale_colour_newRetro}
\alias{scale_colour_newRetro}
\title{scale_colour_newRetro}
\usage{
scale_colour_newRetro(...)
}
\arguments{
\item{...}{pass me the palette}
}
\description{
scale_colour_newRetro
}
