% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_seaPunk.R
\name{scale_colour_seaPunk}
\alias{scale_colour_seaPunk}
\alias{scale_color_seaPunk}
\title{Seapunk Colour and Fill Scales for ggplot2}
\usage{
scale_colour_seaPunk(...)

scale_color_seaPunk(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Seapunk Colour and Fill Scales for ggplot2

Scale Color Seapunk - Alias for Colour
}
\seealso{
\code{\link[ggplot2:scale_colour_discrete]{scale_colour_discrete()}} and 
  \code{\link[ggplot2:scale_fill_discrete]{scale_fill_discrete()}}

\code{\link{scale_colour_seaPunk}}
}
