% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapoRwave.R
\name{scale_colour_vapoRwave}
\alias{scale_colour_vapoRwave}
\title{scale_colour_vapoRwave}
\usage{
scale_colour_vapoRwave(...)
}
\arguments{
\item{...}{pass me the palette}
}
\description{
scale_colour_vapoRwave
}
