% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_avanti.R
\name{scale_fill_avanti}
\alias{scale_fill_avanti}
\title{Avanti Fill Scales for ggplot2}
\usage{
scale_fill_avanti(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Avanti Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_avanti}}
}
