% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cool.R
\name{scale_fill_cool}
\alias{scale_fill_cool}
\title{Cool Fill Scales for ggplot2}
\usage{
scale_fill_cool(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Cool Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_cool}}
}
