% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_crystalPepsi.R
\name{scale_fill_crystalPepsi}
\alias{scale_fill_crystalPepsi}
\title{Crystal Pepsi Fill Scales for ggplot2}
\usage{
scale_fill_crystalPepsi(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Crystal Pepsi Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_crystalPepsi}}
}
