% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_floralShoppe.R
\name{scale_fill_floralShoppe}
\alias{scale_fill_floralShoppe}
\title{Floral Shoppe Fill Scales for ggplot2}
\usage{
scale_fill_floralShoppe(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Floral Shoppe Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_floralShoppe}}
}
