% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hotlineBling.R
\name{scale_fill_hotlineBling}
\alias{scale_fill_hotlineBling}
\title{scale_fill_hotlineBling}
\usage{
scale_fill_hotlineBling(...)
}
\arguments{
\item{...}{pass me the palette}
}
\description{
scale_fill_hotlineBling
}
