% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hyperBubble.R
\name{scale_fill_hyperBubble}
\alias{scale_fill_hyperBubble}
\title{Hyper Bubble Fill Scales for ggplot2}
\usage{
scale_fill_hyperBubble(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Hyper Bubble Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_hyperBubble}}
}
