% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_jazzCup.R
\name{scale_fill_jazzCup}
\alias{scale_fill_jazzCup}
\title{Jazz Cup Fill Scales for ggplot2}
\usage{
scale_fill_jazzCup(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Jazz Cup Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_jazzCup}}
}
