% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_jwz.R
\name{scale_fill_jwz}
\alias{scale_fill_jwz}
\title{jwz Fill Scales for ggplot2}
\usage{
scale_fill_jwz(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
jwz Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_jwz}}
}
