% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_mallSoft.R
\name{scale_fill_mallSoft}
\alias{scale_fill_mallSoft}
\title{Mallsoft Fill Scales for ggplot2}
\usage{
scale_fill_mallSoft(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Mallsoft Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_mallSoft}}
}
