% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_newRetro.R
\name{scale_fill_newRetro}
\alias{scale_fill_newRetro}
\title{New Retro Fill Scales for ggplot2}
\usage{
scale_fill_newRetro(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
New Retro Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_newRetro}}
}
