% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_seaPunk.R
\name{scale_fill_seaPunk}
\alias{scale_fill_seaPunk}
\title{Seapunk Fill Scales for ggplot2}
\usage{
scale_fill_seaPunk(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Seapunk Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_seaPunk}}
}
