% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_sunSet.R
\name{scale_fill_sunSet}
\alias{scale_fill_sunSet}
\title{Sunset Fill Scales for ggplot2}
\usage{
scale_fill_sunSet(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Sunset Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_sunSet}}
}
