% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_vapoRwave.R
\name{scale_fill_vapoRwave}
\alias{scale_fill_vapoRwave}
\title{VapoRwave Fill Scales for ggplot2}
\usage{
scale_fill_vapoRwave(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
VapoRwave Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_vapoRwave}}
}
