% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_seaPunk.R
\name{seaPunk_pal}
\alias{seaPunk_pal}
\title{Seapunk Colour Palette}
\usage{
seaPunk_pal()
}
\value{
A character vector of color hex codes representing the Seapunk palette.
}
\description{
Colour scales inspired by the Seapunk subculture that emerged on Tumblr in 2011,
characterized by aquatic-themed fashion, 3D net art, and 90s pop culture references.
}
