% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_sunSet.R
\name{sunSet_pal}
\alias{sunSet_pal}
\title{Sunset Colour Palette}
\usage{
sunSet_pal()
}
\value{
A character vector of color hex codes representing the Sunset palette.
}
\description{
Colour scales inspired by the vibrant hues of a sunset.
}
