% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_vapoRwave.R
\name{vapoRwave_pal}
\alias{vapoRwave_pal}
\title{VapoRwave Colour Palette}
\usage{
vapoRwave_pal()
}
\value{
A character vector of color hex codes representing the vapoRwave palette.
}
\description{
Colour scales inspired by the vapoRwave aesthetic.
}
