% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crystalPepsi.R
\name{scale_colour_crystalPepsi}
\alias{scale_colour_crystalPepsi}
\title{scale_colour_crystalPepsi}
\usage{
scale_colour_crystalPepsi(...)
}
\arguments{
\item{...}{pass me the palette}
}
\description{
scale_colour_crystalPepsi
}
