% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_crystalPepsi.R
\name{scale_colour_crystalPepsi}
\alias{scale_colour_crystalPepsi}
\alias{scale_color_crystalPepsi}
\title{Crystal Pepsi Colour and Fill Scales for ggplot2}
\usage{
scale_colour_crystalPepsi(...)

scale_color_crystalPepsi(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Colour scales based on the Crystal Pepsi craze of the early 1990's.
}
\seealso{
\code{\link[ggplot2:scale_colour_discrete]{scale_colour_discrete()}} and 
  \code{\link[ggplot2:scale_fill_discrete]{scale_fill_discrete()}}

\code{\link{scale_colour_crystalPepsi}}
}
