% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hotlineBling.R
\name{scale_colour_hotlineBling}
\alias{scale_colour_hotlineBling}
\alias{scale_color_hotlineBling}
\title{Hotline Bling Colour and Fill Scales for ggplot2}
\usage{
scale_colour_hotlineBling(...)

scale_color_hotlineBling(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Colour scales inspired by Drake's "Hotline Bling" music video.
}
\seealso{
\code{\link[ggplot2:scale_colour_discrete]{scale_colour_discrete()}} and 
  \code{\link[ggplot2:scale_fill_discrete]{scale_fill_discrete()}}

\code{\link{scale_colour_hotlineBling}}
}
