% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hyperBubble.R
\name{scale_colour_hyperBubble}
\alias{scale_colour_hyperBubble}
\alias{scale_color_hyperBubble}
\title{Hyper Bubble Colour and Fill Scales for ggplot2}
\usage{
scale_colour_hyperBubble(...)

scale_color_hyperBubble(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Colour scales inspired by the vibrant and dynamic aesthetics of hyperpop music.
}
\seealso{
\code{\link[ggplot2:scale_colour_discrete]{scale_colour_discrete()}} and 
  \code{\link[ggplot2:scale_fill_discrete]{scale_fill_discrete()}}

\code{\link{scale_colour_hyperBubble}}
}
