% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_mallSoft.R
\name{scale_colour_mallSoft}
\alias{scale_colour_mallSoft}
\title{scale_colour_mallSoft}
\usage{
scale_colour_mallSoft(...)
}
\arguments{
\item{...}{pass me the palette}
}
\description{
scale_colour_mallSoft
}
