% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_newRetro.R
\name{scale_colour_newRetro}
\alias{scale_colour_newRetro}
\alias{scale_color_newRetro}
\title{New Retro Colour and Fill Scales for ggplot2}
\usage{
scale_colour_newRetro(...)

scale_color_newRetro(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
New Retro Colour and Fill Scales for ggplot2

Scale Color New Retro - Alias for Colour
}
\seealso{
\code{\link[ggplot2:scale_colour_discrete]{scale_colour_discrete()}} and 
  \code{\link[ggplot2:scale_fill_discrete]{scale_fill_discrete()}}

\code{\link{scale_colour_newRetro}}
}
