% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_crystalPepsi.R
\name{scale_fill_crystalPepsi}
\alias{scale_fill_crystalPepsi}
\title{crystalPepsi fill}
\usage{
scale_fill_crystalPepsi(...)
}
\arguments{
\item{...}{pass the palette}
}
\description{
crystalPepsi fill
}
