% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_floralShoppe.R
\name{scale_fill_floralShoppe}
\alias{scale_fill_floralShoppe}
\title{scale_fill_floralShoppe}
\usage{
scale_fill_floralShoppe(...)
}
\arguments{
\item{...}{pass me the palette}
}
\description{
scale_fill_floralShoppe
}
