% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hotlineBling.R
\name{scale_fill_hotlineBling}
\alias{scale_fill_hotlineBling}
\title{Hotline Bling Fill Scales for ggplot2}
\usage{
scale_fill_hotlineBling(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{discrete_scale}}}
}
\description{
Hotline Bling Fill Scales for ggplot2
}
\seealso{
\code{\link{scale_colour_hotlineBling}}
}
