% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hyperBubble.R
\name{scale_fill_hyperBubble}
\alias{scale_fill_hyperBubble}
\title{scale_fill_hyperBubble}
\usage{
scale_fill_hyperBubble(...)
}
\arguments{
\item{...}{pass me the palette}
}
\description{
scale_fill_hyperBubble
}
